Add python test case, update makefile and minor formatting changes

+++ quaternion/src/quaternion.h
@@ -1,27 +1,27 @@
 #ifndef _QUATERNION_H
 #define _QUATERNION_H
 
+#include <math.h>
 #include <stdbool.h>
 #include <stdint.h>
-#include <math.h>
 
 #ifndef M_PI
 #define M_PI 3.14159265
 #endif
 
 typedef struct Quaternion {
-  double w;
-  double x;
-  double y;
-  double z;
-  void (*init)(struct Quaternion*);
-  void (*normalise)(struct Quaternion*);
-  void (*fromEuler)(struct Quaternion*, double, double, double);
+  float w;
+  float x;
+  float y;
+  float z;
+  void (*init)(struct Quaternion *);
+  void (*normalise)(struct Quaternion *);
+  void (*fromEuler)(struct Quaternion *, float, float, float);
 } Quaternion;
 
-Quaternion Quaternion_mul(Quaternion*, Quaternion*);
-void Quaternion_init(Quaternion*);
-void Quaternion_normalise(Quaternion*);
-void Quaternion_fromEuler(Quaternion*, double, double, double);
+Quaternion Quaternion_mul(Quaternion *, Quaternion *);
+void Quaternion_init(Quaternion *);
+void Quaternion_normalise(Quaternion *);
+void Quaternion_fromEuler(Quaternion *, float, float, float);
 
 #endif
